requirements.txt
discord.py
pytube
python-dotenv
gofile2
imgurpython
